fix use-after-deletion of CPlayBoard
Previous attempt to create RAII UniqueGeoBoard wrapper tried to
hold a pointer to the associated containing object.  However,
on move operations, copying that pointer meant copied object
held a pointer to deleted containing object, so that doesn't
work.  Fortunately, CGeomorphicBoard already contains a pointer
to the needed object, so move CPlayBoard code that needed the
pointer to CGeomorphicBoard.