Display geoboard CellNumbers as subunit intends
instead of relative to the overall board.  I consider this an
improvement since rulebooks always use the hex numbers printed
on the map, which can't change in the physical world when the
maps are rearranged.  This might also work well for multi-map
games where there are small scenarios on a single map and large
scenarios that use multiple maps.