LBoxTray:  only show tooltip if its side is visible
Before this commit, LBoxTray was using different code to decide
whether to show the inactive side of a piece than to decide
whether to show the side's tooltip.  This resulted in showing
tooltips for sides that were not shown, which doesn't make sense.
Therefore, factor out the code that decided whether to show
inactive sides so the tooltip can use the same code.