treat random side choice as hidden move
When pulling from a tray that returns pieces flipped to a random
side, the flip must be recorded as a move so that CB always knows
the piece's active side.  However, during playback, there's no
reason to display the side flip as separate from moving the piece
from the tray to the board.