CPieceObj:  replace m_rctExtent with center point
The size of CPieceObj depends on CPieceTable, so have CPieceObj
store its center point, not the entire m_rctExtent, and then
look up the size when the rect is needed.

Previous attempts to solve problems related to keeping
CPieceObj::m_rctExtent up-to-date never fully succeeded, so try
to solve the problem by calculating rect extent when needed.