VwTilesl:  better enforce that VwTilesl is never the active view
The previous code in CTileSelView::OnLButtonDown() appeared to
be intended to make CBitEditView the active view whenever
CTileSelView would have become active, but it was incomplete.
(If the user right-clicked the CTileSelView, it became active.)
So modify the code to be more general about relaying the active
status to CBitEditView.